+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Added
 - `Schema\Grammar::compileDropForeign` to allow dropping foreign key constraints (#163)
 - `Schema\Builder::dropAllTables` works properly, dropping foreign keys, indexes, then tables in order of interleaving (#161)
 - Support for inserting and selecting array of DateTime/Numeric objects (#168)
+- Allow pretending for DDL statements (#170) 
 
 Changed
 - `Query\Builder::lock()` no longer throw an error and will be ignored instead (#156)